move validator to region submodule
